change to home_url to make browser reload work in more local settings
In some cases (as mine) "/wordpress/" may occur in the URL. This can arise when WordPress is installed in a directory other than public_html (or whatever web root your host uses). This can be useful if you're trying to achieve that end. If instead, you are trying to reliably access the URL of the home page of the site, be sure to use home_url().